worker/requirements.txt
mysql.connector # Para conectar ao mySQL
pika          # Para conectar ao RabbitMQ
jinja2        # Para templates HTML no Worker
pdfkit        # Para gerar PDFs a partir de HTML (no Worker)
